Module OUnitConf

exception Parse_error of string

type conf = OUnitPropList.t

type 'a var = conf -> 'a

type metadata = {

help : string;

get_print : conf -> string;

parse_set : string -> conf -> unit;

cli : conf -> (string * Stdlib.Arg.spec * string) list;

}

val metaconf : (string, metadata) Stdlib.Hashtbl.t

val check_variable_name : string -> unit

val cli_name : string -> string

val subst : conf -> (string * string) list -> string -> string

val make : name:string -> parse:(string -> 'a) -> print:('a0 -> string) -> default:'a1 -> help:string -> fcli:('a2 -> ('a3 -> unit) -> (string * Stdlib.Arg.spec * string) list) -> unit -> 'a4 var

val make_string : string -> string -> string -> string var

val make_string_subst : string -> string -> string -> ?extra_subst:(string * string) list -> conf -> string

val make_string_opt : string -> string option -> string -> string option var

val make_string_subst_opt : string -> string option -> string -> ?extra_subst:(string * string) list -> conf -> string option

val make_int : string -> int -> string -> int var

val make_float : string -> float -> string -> float var

val make_bool : string -> bool -> string -> bool var

val make_enum : string -> (unit -> (string * 'a) list) -> string -> string -> conf -> string * 'b

val make_exec : string -> string var

val set : origin:string -> conf -> string -> string -> unit

val file_parse : conf -> string -> unit

val env_parse : conf -> unit

val cli_parse : ?argv:string array -> (Stdlib.Arg.key * Stdlib.Arg.spec * Stdlib.Arg.doc) list -> conf -> unit

val default : ?preset:(string * string) list -> unit -> (int, unit -> unit) Stdlib.Hashtbl.t

val load : ?preset:(string * string) list -> ?argv:string array -> (Stdlib.Arg.key * Stdlib.Arg.spec * Stdlib.Arg.doc) list -> (int, unit -> unit) Stdlib.Hashtbl.t

Load test options from file, environment and command line (in this order). Not that extra_specs is here for historical reason, better use make to create command line options.

val dump : conf -> (string * string) list