Module OUnitRunner

val shards : int OUnitConf.var

Number of shards to use. The way the shards are used depends on the type of runner.

val run_gc_full_major : bool OUnitConf.var

Whether or not run a Gc.full_major in between tests. This adds time when running tests, but helps to avoid unexpected error due to finalisation of values allocated during a test.

val run_one_test : OUnitConf.conf -> OUnitTest.logger -> OUnitShared.shared -> OUnitTest.path -> OUnitTest.test_fun -> OUnitTest.result_full * OUnitTest.result_full list

Common utilities to run test.

type runner = OUnitConf.conf -> OUnitTest.logger -> OUnitChooser.chooser -> (OUnitTest.path * OUnitTest.test_length * OUnitTest.test_fun) list -> OUnitTest.result_list

val sequential_runner : runner

val all : (int * (OUnitPlugin.name * runner)) list Stdlib.ref

val register : OUnitPlugin.name -> int -> runner -> unit

val of_name : OUnitPlugin.name -> runner

val choice : OUnitConf.conf -> OUnitPlugin.name * runner

val preset : (OUnitPlugin.name * OUnitPlugin.name) list -> (OUnitPlugin.name * OUnitPlugin.name) list

module GenericWorker : sig ... end

Build worker based runner.